backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
plaid-python==17.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pyotp==2.9.0
qrcode[pil]==7.4.2
reportlab==4.2.0
pandas==2.2.2
httpx==0.27.0
python-dotenv==1.0.1
slowapi==0.1.9
pydantic[email]==2.7.1
pydantic-settings==2.3.4
psycopg2-binary==2.9.9
pytest==8.2.0
pytest-asyncio==0.23.7
